Split model tests into smaller files.
